ODS (OpenDocument Spreadsheet) Exporter for EMF (c.d.)
 - Potential fix for performance issue - over 7 times (700+%) speed-up

 - Refactoring

 - (temporarily) large-data set (trees) for performance measurement
