+++ org.gecko.emf.exporter.ods.api/.gitignore
@@ -1,0 +1,2 @@
+/bin/
+/bin_test/
