+++ org.gecko.emf.exporter.ods.api/src/org/gecko/emf/exporter/ods/api/package-info.java
@@ -1,0 +1,3 @@
+@org.osgi.annotation.versioning.Version("1.0.0")
+@org.osgi.annotation.bundle.Export
+package org.gecko.emf.exporter.ods.api;
