+++ org.gecko.emf.exporter.ods.tests/test.bndrun
@@ -5,7 +5,25 @@
 -resolve.effective: active
 
 -runbundles.junit5: ${test.runbundles}
-	
+
+-runrequires: \
+	bnd.identity;id='org.gecko.emf.exporter.ods',\
+	bnd.identity;id='org.gecko.emf.exporter.ods.tests'
+
+-runee: JavaSE-11
+
+-runtrace: true
+
+-runproperties.debug: \
+	felix.log.level=4,\
+	org.osgi.service.log.admin.loglevel=DEBUG,\
+
+-runproperties.poi: \
+	org.osgi.framework.bootdelegation=org.w3c.dom
+
+-runproperties.trees: \
+	TREES_DATASET_XMI="${.}/testdata/Trees/CityTree.xmi"
+
 -runbundles: \
 	junit-jupiter-api;version='[5.8.2,5.8.3)',\
 	junit-platform-commons;version='[1.8.2,1.8.3)',\
@@ -47,16 +65,11 @@
 	org.apache.commons.logging;version='[1.2.0,1.2.1)',\
 	slf4j.api;version='[1.7.36,1.7.37)',\
 	slf4j.simple;version='[1.7.25,1.7.26)',\
-	org.gecko.emf.bson;version=snapshot,\
-	org.gecko.emf.collections;version=snapshot
-
--runrequires: bnd.identity;id='org.gecko.emf.exporter.ods.tests'
-
--runee: JavaSE-11
-
--runtrace: true
-
--runproperties.debug: \
-	felix.log.level=4,\
-	org.osgi.service.log.admin.loglevel=DEBUG
-
+	com.google.guava;version='[32.1.2,32.1.3)',\
+	com.google.guava.failureaccess;version='[1.0.1,1.0.2)',\
+	org.gecko.emf.trees.model;version=snapshot,\
+	org.osgi.service.component;version='[1.4.0,1.4.1)',\
+	org.apache.commons.commons-compress;version='[1.21.0,1.21.1)',\
+	org.apache.servicemix.bundles.poi;version='[5.2.2,5.2.3)',\
+	org.gecko.emf.collections;version=snapshot,\
+	org.gecko.emf.ods;version=snapshot
