--- org.gecko.emf.exporter.ods/src/org/gecko/emf/exporter/ods/EMFODSExportOptions.java
@@ -1,23 +1,0 @@
-/**
- * Copyright (c) 2012 - 2023 Data In Motion and others.
- * All rights reserved. 
- * 
- * This program and the accompanying materials are made available under the terms of the 
- * Eclipse Public License v2.0 which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v20.html
- * 
- * Contributors:
- *     Data In Motion - initial API and implementation
- */
-package org.gecko.emf.exporter.ods;
-
-import org.gecko.emf.exporter.EMFExportOptions;
-
-/**
- * TODO: add javadoc
- * @author michal
- * @since Aug 29, 2023
- */
-public interface EMFODSExportOptions extends EMFExportOptions {
-	// TODO: extract from 'org.gecko.emf.exporter.EMFExportOptions' ODS format-specific options
-}
