+++ org.gecko.emf.exporter/src/org/gecko/emf/exporter/EMFExportOptions.java
@@ -27,17 +27,6 @@ public interface EMFExportOptions {
 	// extract and export metadata as additional sheets
 	String OPTION_EXPORT_METADATA = "EXPORT_METADATA";
 
-	// automatically adjust column width based on contents
-	String OPTION_ADJUST_COLUMN_WIDTH = "ADJUST_COLUMN_WIDTH"; // TODO: extract to org.gecko.emf.exporter.ods.EMFODSExportOptions
-
-	// generate links for references
-	String OPTION_GENERATE_LINKS = "GENERATE_LINKS"; // TODO: extract to org.gecko.emf.exporter.ods.EMFODSExportOptions
-	
 	// generate mapping table
 	String OPTION_ADD_MAPPING_TABLE = "ADD_MAPPING_TABLE";
-	
-	// TODO: freezing rows is currently not supported in SODS
-	// freeze header row
-//	String OPTION_FREEZE_HEADER_ROW = "FREEZE_HEADER_ROW";
-
 }
