+++ org.gecko.emf.ods.tests/bnd.bnd
@@ -12,4 +12,5 @@ Bundle-Version: 1.0.0.SNAPSHOT
 	org.eclipse.emf.ecore,\
 	org.gecko.emf.osgi.example.model.basic,\
 	org.gecko.emf.exporter;version=latest,\
+	org.gecko.emf.exporter.ods.api;version=latest,\
 	org.apache.commons.commons-text
