+++ org.gecko.emf.trees.model/.gitignore
@@ -1,0 +1,1 @@
+/bin/
