+++ org.gecko.emf.trees.model/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,6 @@
+eclipse.preferences.version=1
+encoding//.settings/org.eclipse.core.resources.prefs=UTF-8
+encoding//.settings/org.eclipse.jdt.core.prefs=UTF-8
+encoding//.settings/org.eclipse.jdt.ui.prefs=UTF-8
+encoding//model/trees.genmodel=UTF-8
+encoding/bnd.bnd=UTF-8
