+++ org.gecko.emf.trees.model/bnd.bnd
@@ -1,0 +1,25 @@
+# sets the usually required buildpath using the bnd library, you can extend it with the normal -buildpath to your liking
+src src,src-gen
+
+# The code generation takes a bit of time and makes the build a bit slower.
+# It might be a good idea to put comments around it, when you don't need it
+-generate:\
+	model/trees.genmodel;\
+		generate=geckoEMF;\
+		genmodel=model/trees.genmodel;\
+		output=src-gen
+# Add this attribute to find some logging information
+#		logfile=test.log;\
+
+# If this is not done with a instruction namespace, bnd would think you like
+# to take full control over whats exported and imported and will ignore @Export
+# Bundle Annotations		
+-includeresource.model: model=model
+
+-library: enable-emf
+
+Bundle-Version: 1.0.0.SNAPSHOT
+Export-Package: \
+	trees,\
+	trees.impl,\
+	trees.util
